Moved IPC sink test to correct location.
