+++ src/io/ipc/write/stream_async.rs
@@ -259,54 +259,3 @@ where
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::StreamSink;
-    use crate::{
-        array::{Array, Float32Array, Int32Array},
-        chunk::Chunk,
-        datatypes::{DataType, Field, Schema},
-        io::ipc::read::stream_async::{read_stream_metadata_async, AsyncStreamReader},
-    };
-    use futures::{SinkExt, TryStreamExt};
-    use std::sync::Arc;
-
-    // Verify round trip data integrity when using async read + write.
-    #[test]
-    fn test_stream_async_roundtrip() {
-        futures::executor::block_on(async move {
-            let mut data = vec![];
-            for i in 0..5 {
-                let a1 = Int32Array::from(&[Some(i), None, Some(i + 1)]);
-                let a2 = Float32Array::from(&[None, Some(i as f32), None]);
-                let chunk = Chunk::new(vec![
-                    Arc::new(a1) as Arc<dyn Array>,
-                    Arc::new(a2) as Arc<dyn Array>,
-                ]);
-                data.push(chunk);
-            }
-            let schema = Schema::from(vec![
-                Field::new("a1", DataType::Int32, true),
-                Field::new("a2", DataType::Float32, true),
-            ]);
-
-            let mut buffer = vec![];
-            let mut sink = StreamSink::new(&mut buffer, schema.clone(), None, Default::default());
-            for chunk in &data {
-                sink.feed(chunk.clone()).await.unwrap();
-            }
-            sink.close().await.unwrap();
-            drop(sink);
-
-            let mut reader = &buffer[..];
-            let metadata = read_stream_metadata_async(&mut reader).await.unwrap();
-            assert_eq!(schema, metadata.schema);
-            let stream = AsyncStreamReader::new(reader, metadata);
-            let out = stream.try_collect::<Vec<_>>().await.unwrap();
-            for i in 0..5 {
-                assert_eq!(data[i], out[i]);
-            }
-        })
-    }
-}
