+++ tests/it/io/ipc/read_stream_async.rs
@@ -1,9 +1,15 @@
-use futures::StreamExt;
+use std::sync::Arc;
+
+use futures::{SinkExt, StreamExt, TryStreamExt};
 use tokio::fs::File;
 use tokio_util::compat::*;
 
+use arrow2::array::{Array, Float32Array, Int32Array};
+use arrow2::chunk::Chunk;
+use arrow2::datatypes::{DataType, Field, Schema};
 use arrow2::error::Result;
 use arrow2::io::ipc::read::stream_async::*;
+use arrow2::io::ipc::write::stream_async::StreamSink;
 
 use crate::io::ipc::common::read_gzip_json;
 
@@ -42,4 +48,39 @@ async fn test_file(version: &str, file_name: &str) -> Result<()> {
 #[tokio::test]
 async fn write_async() -> Result<()> {
     test_file("1.0.0-littleendian", "generated_primitive").await
+}
+
+#[tokio::test]
+async fn test_stream_async_roundtrip() {
+    let mut data = vec![];
+    for i in 0..5 {
+        let a1 = Int32Array::from(&[Some(i), None, Some(i + 1)]);
+        let a2 = Float32Array::from(&[None, Some(i as f32), None]);
+        let chunk = Chunk::new(vec![
+            Arc::new(a1) as Arc<dyn Array>,
+            Arc::new(a2) as Arc<dyn Array>,
+        ]);
+        data.push(chunk);
+    }
+    let schema = Schema::from(vec![
+        Field::new("a1", DataType::Int32, true),
+        Field::new("a2", DataType::Float32, true),
+    ]);
+
+    let mut buffer = vec![];
+    let mut sink = StreamSink::new(&mut buffer, schema.clone(), None, Default::default());
+    for chunk in &data {
+        sink.feed(chunk.clone()).await.unwrap();
+    }
+    sink.close().await.unwrap();
+    drop(sink);
+
+    let mut reader = &buffer[..];
+    let metadata = read_stream_metadata_async(&mut reader).await.unwrap();
+    assert_eq!(schema, metadata.schema);
+    let stream = AsyncStreamReader::new(reader, metadata);
+    let out = stream.try_collect::<Vec<_>>().await.unwrap();
+    for i in 0..5 {
+        assert_eq!(data[i], out[i]);
+    }
 }
